refactor(tests): split useInsightEnrichment.test.ts into 3 focused files
Splits the 1977-line monolith into:
- useInsightEnrichment-ws: mergeEnrichments (empty/enriched), agent not connected, agent unavailable, WebSocket connection, exponential backoff retry
- useInsightEnrichment-http: HTTP enrichment request, return value shape, cache TTL and hash dedup, multiple enrichments overwrite, severity edge cases, remediation fallback
- useInsightEnrichment-edge: HTTP error handling, debounce behavior, optional payload fields, WebSocket edge cases, cleanup on unmount, concurrent hook instances, requestEnrichment guards, applyEnrichments edge cases, WS reconnect, severity undefined edge case, hashInsights determinism

Fixes #8624
